requirements.txt
torch 1.3.1
torchvision 0.4.2
h5py 2.10.0
Pillow 6.2.1
pandas 1.0.2
scikit-image  0.17.2
transforms3d 0.3.1
tensorboardX 2.0
scikit-learn 0.22.1
